Removed isValidCredentials service, fixed the calls to the services
Changed the client from the custom version with the isValidCredentials
to the old one with just the search and update methods

Added the dependencies to the repo

Removed the .jsp client

In SPSElite.java 

Moved the calls to the configuration data to the method void
setConnectionParameters()

Added calls to all the methods that use the proxy

Added more debugging to the code

testConnection() - it now uses the search() service instead of the now
deleted isValidUserCredentials()

discoverSchema() - Changed the type of the isActive attribute from
boolean to String

createUpdate() - 
Fixed an error were the method will not validate if the transaction was
made

userExist() -
Added a call to the setConnectionParameters() method to allow the
service to retrieve the connector configuration 

Fixed a NPE caused by the missing parameter of the searchRequestCriteria
object.


loadUsers() -
Fixed a NPE caused by parameter missing to the searchRequestCriteria
object.